# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitrack",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/routines/{user_id}": {
            "get": {
                "tags": [
                    "Routine"
                ],
                "summary": "Obtener las rutinas por ID de usuario",
                "description": "Devuelve una lista de rutinas asociadas a un usuario específico.",
                "operationId": "getRoutinesByUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de rutinas del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Routine"
                ],
                "summary": "Insertar una nueva rutina para un usuario",
                "description": "Crea una nueva rutina personalizada asociada a un usuario.",
                "operationId": "insertRoutineByUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "imageUri",
                                    "exerciseIds"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rutina de pierna"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Ejercicios centrados en piernas y glúteos"
                                    },
                                    "imageUri": {
                                        "type": "string",
                                        "example": "https://miapp.com/images/pierna.png"
                                    },
                                    "exerciseIds": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            101,
                                            102,
                                            103
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rutina insertada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Routine insertada exitosamente"
                                        },
                                        "routine": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "La rutina ya existe para ese usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Routine ya existe para ese usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/routines/{user_id}/{routine_id}": {
            "delete": {
                "tags": [
                    "Routine"
                ],
                "summary": "Eliminar una rutina por ID de usuario e ID de rutina",
                "description": "Elimina una rutina específica de un usuario dado.",
                "operationId": "deleteRoutineByUserIdAndRoutineId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "routine_id",
                        "in": "path",
                        "description": "ID de la rutina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rutina eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Routine eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rutina no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Routine no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token en formato Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}